pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "ai-privacy-anonymizer"
version = "0.1.0"
description = "Local-first privacy anonymizer for Italian PII."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Sergio Dogliani" }]
dependencies = []

[project.optional-dependencies]
dev = ["pytest>=8.0"]
rich = ["rich>=13.0"]
office = [
  "python-docx>=1.1",
  "openpyxl>=3.1",
  "python-pptx>=0.6",
]
ml = [
  "gliner>=0.2.5",
]
documents = [
  "pypdf>=4.0",
  "reportlab>=4.0",
  "pillow>=10.0",
  "pytesseract>=0.3.10",
  "striprtf>=0.0.26",
]
docling = [
  "docling>=2.0",
  "docling-core>=2.0",
]
webui = [
  "gradio>=4.0",
]
api = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "python-multipart>=0.0.9",
]
all = [
  "python-docx>=1.1",
  "openpyxl>=3.1",
  "python-pptx>=0.6",
  "pypdf>=4.0",
  "reportlab>=4.0",
  "pillow>=10.0",
  "pytesseract>=0.3.10",
  "striprtf>=0.0.26",
  "docling>=2.0",
  "docling-core>=2.0",
  "gradio>=4.0",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "python-multipart>=0.0.9",
  "rich>=13.0",
  "gliner>=0.2.5",
]

[project.scripts]
privacy-anonymizer = "privacy_anonymizer.cli:main"
privacy-anonymizer-web = "privacy_anonymizer.webui:launch"

[tool.hatch.build.targets.wheel]
packages = ["src/privacy_anonymizer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
